add a "helper" and a "hooks" class. Move constants needed by multiple classes to a "constants" class
